refactor(upstream): extract pool/connect.rs + pool/connect_stdio.rs (lab-kvji.12.4 step 6)
Move HTTP/WebSocket connectors (connect_upstream, connect_http_upstream,
connect_websocket_upstream, stable_jitter_seed, log_oauth_uncapped_once,
runtime_origin_label) into connect.rs with the 3 oauth-connect tests; move stdio +
in-process connectors (connect_stdio_upstream, connect_in_process_service_peer) into
connect_stdio.rs. All pub(super). NO-TOUCH: stdio process-group guard arm/disarm
ordering moved byte-identical. Behavior unchanged. 1423 green.

Bead: lab-kvji.12.4
